requirements.txt
flask
flask_restful
tensorflow
pillow
numpy